Shared finding ownerships
