+++ lib/library_version_analysis/npm.rb
@@ -1,5 +1,9 @@
+require "library_version_analysis/ownership"
+
 module LibraryVersionAnalysis
   class Npm
+    include LibraryVersionAnalysis::Ownership
+
     def initialize(github_repo)
       @github_repo = github_repo
     end
@@ -207,34 +211,6 @@ module LibraryVersionAnalysis
       else
         add_transitive_ownerships(parsed_results)
       end
-    end
-
-    def add_transitive_ownerships(parsed_results)
-      parsed_results.select { |_, result_data| LibraryVersionAnalysis::CheckVersionStatus.unknown_owner?(result_data.owner) }.each do |name, line_data|
-        @current_library = name
-        owner = find_owner(line_data.dependency_graph, parsed_results)
-        line_data.owner = LibraryVersionAnalysis::CheckVersionStatus.unknown_owner?(owner) ? :unknown : owner
-      end
-    end
-
-    def find_owner(dependency_graph, parsed_results)
-      return nil if dependency_graph.nil?
-
-      owner = parsed_results[dependency_graph.name]&.owner
-      return owner unless LibraryVersionAnalysis::CheckVersionStatus.unknown_owner?(owner)
-
-      parent_owner = nil
-
-      dependency_graph.parents&.each do |parent|
-        parent_owner = parsed_results[parent.name]&.owner
-        return parent_owner unless LibraryVersionAnalysis::CheckVersionStatus.unknown_owner?(parent_owner)
-
-        parent_owner = find_owner(parent, parsed_results)
-        is_unknown = LibraryVersionAnalysis::CheckVersionStatus.unknown_owner?(parent_owner)
-        break unless is_unknown
-      end
-
-      return parent_owner
     end
 
     def add_transitive_ownerships_legacy(parsed_results)
